+++ test/Console.t.sol
@@ -7,7 +7,7 @@ contract ConsoleTest is Test {
     function testLogSomething() public {
         console.log("Log something here", 123);
 
-        int x = -1;
+        int256 x = -1;
         console.logInt(x);
     }
 }
