+++ test/Time.t.sol
@@ -6,7 +6,7 @@ import {Auction} from "../src/Time.sol";
 
 contract TimeTest is Test {
     Auction public auction;
-    uint public startAt;
+    uint256 public startAt;
 
     function setUp() public {
         auction = new Auction();
@@ -30,7 +30,7 @@ contract TimeTest is Test {
     }
 
     function testTimestamp() public {
-        uint t = block.timestamp;
+        uint256 t = block.timestamp;
         skip(100);
         assertEq(block.timestamp, t + 100);
 
@@ -39,7 +39,7 @@ contract TimeTest is Test {
     }
 
     function testBlockNumber() public {
-        uint b = block.number;
+        uint256 b = block.number;
         vm.roll(999);
         assertEq(block.number, 999);
     }
